[s1718] Implement some server side summons spells Thanks cyberium
Add spells: 34810, 34817, 34818, 34819, 35153, 35904, 35905, 35906

Signed-off-by: Schmoozerd <schmoozerd@scriptdev2.com>
[s1718] Change behaviour of text selection for gossip menus

This way the behaviour will be well defined in case multiple conditions match.
Now the text with the highest condition-id will be taken.
[s1719] Remove unused argument from OutdoorPvP::HandlePlayerKillInsideArea
[s1720] Remove unneeded InBattleGround() as it is already checked by GetBattleGround()

Also call OutdoorPvP handlers before instance script ones
[s1721] Add HandleEvent and HandleGameObjectCreate to battleground class
[s1722] Prevent saving/loading/init capture points in instances

Reason for this are:
1) We don't want them saved/loaded between bg games
2) Saving cp slider is done by go id and would not work for instances
3) There are visual-only capture points in bg (eots)
4) It would be possible to only enable the capture points on bg start

TODO: Move this to OutdoorPvP base class
[s1723] Merge Player::CanUseOutdoorCapturePoint into Player::CanUseCapturePoint
[s1724] Add support to link summoned npcs with following

Thanks to crackm for insisting that this needs a core change and testing
Fixup script_command documentation. Thanks to crackm for pointing
Add EVENT_T_TIMER to EventAI. This is a generic timer that ticks out of combat and in combat

Signed-off-by: Schmoozerd <schmoozerd@cmangos>
Update src/game/CreatureEventAI.h
[s1725] Ensure EVENT_T_GENERIC_TIMER is proper initialized on create. Thanks to Reamer for spotting this glitch
Forgoten change from last commit [s1703]

Signed-off-by: Salja <salja2012@hotmail.de>